+++ x-pack/functionbeat/_meta/beat.yml
@@ -263,9 +263,9 @@ functionbeat.provider.gcp.functions:
     # Description of the method to help identify them when you run multiples functions.
     description: "Google Cloud Function for Pub/Sub"
 
-    # The maximum memory allocated for this function, the configured size must be a factor of 64.
-    # Default is 256MiB.
-    #memory_size: 256MiB
+    # The maximum memory allocated for this function.
+    # Default is 256MB.
+    #memory_size: 256MB
 
     # Execution timeout in seconds. If the function does not finish in time,
     # it is considered failed and terminated. Default is 60s.
@@ -302,9 +302,9 @@ functionbeat.provider.gcp.functions:
     # Description of the method to help identify them when you run multiples functions.
     description: "Google Cloud Function for Cloud Storage"
 
-    # The maximum memory allocated for this function, the configured size must be a factor of 64.
-    # Default is 256MiB.
-    #memory_size: 256MiB
+    # The maximum memory allocated for this function.
+    # Default is 256MB.
+    #memory_size: 256MB
 
     # Execution timeout in seconds. If the function does not finish in time,
     # it is considered failed and terminated. Default is 60s.
