+++ x-pack/functionbeat/config/config.go
@@ -10,13 +10,9 @@ package config
 import (
 	"fmt"
 	"regexp"
-	"unicode"
-
-	humanize "github.com/dustin/go-humanize"
 
 	"github.com/elastic/beats/libbeat/cfgfile"
 	"github.com/elastic/beats/libbeat/common"
-	"github.com/elastic/beats/libbeat/common/cfgwarn"
 )
 
 var (
@@ -137,41 +133,4 @@ func (p *ProviderConfig) Validate() error {
 		names[fc.Name] = true
 	}
 	return nil
-}
-
-// MemSizeFactor64 implements a human understandable format for bytes but also make sure that all
-// values used are a factor of 64.
-type MemSizeFactor64 int
-
-// Unpack converts a size defined from a human readable format into bytes and verifies that the value
-// is a multiple of 64. If the value is not multple of 64, it returns an error.
-func (m *MemSizeFactor64) Unpack(v string) error {
-	sz, err := humanize.ParseBytes(v)
-	if isRawBytes(v) {
-		cfgwarn.Deprecate("7.0.0", "size now requires a unit (KiB, MiB, etc...), current value: %s.", v)
-	}
-	if err != nil {
-		return err
-	}
-
-	if sz%64 != 0 {
-		return fmt.Errorf("number is not a multiple of 64, %d bytes (user value: %s)", sz, v)
-	}
-
-	*m = MemSizeFactor64(sz)
-	return nil
-}
-
-// Megabytes return the value in megabytes.
-func (m *MemSizeFactor64) Megabytes() int {
-	return int(*m) / 1024 / 1024
-}
-
-func isRawBytes(v string) bool {
-	for _, c := range v {
-		if !unicode.IsDigit(c) {
-			return false
-		}
-	}
-	return true
 }
