+++ x-pack/functionbeat/config/config_test.go
@@ -8,7 +8,6 @@
 package config
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -110,40 +109,5 @@ func TestFunctionName(t *testing.T) {
 		f := functionName("")
 		err := f.Unpack("hello world")
 		assert.Error(t, err)
-	})
-}
-
-func TestMemSizeFactor64(t *testing.T) {
-	t.Run("human format", func(t *testing.T) {
-		t.Run("value is a factor of 64", func(t *testing.T) {
-			v := MemSizeFactor64(0)
-			err := v.Unpack("128MiB")
-			if !assert.NoError(t, err) {
-				return
-			}
-			assert.Equal(t, MemSizeFactor64(128*1024*1024), v)
-		})
-	})
-
-	t.Run("raw value", func(t *testing.T) {
-		t.Run("value is a factor of 64", func(t *testing.T) {
-			v := MemSizeFactor64(0)
-			err := v.Unpack(fmt.Sprintf("%d", 128*1024*1024))
-			if !assert.NoError(t, err) {
-				return
-			}
-			assert.Equal(t, MemSizeFactor64(128*1024*1024), v)
-		})
-
-		t.Run("value is not a factor of 64", func(t *testing.T) {
-			v := MemSizeFactor64(0)
-			err := v.Unpack("121")
-			assert.Error(t, err)
-		})
-	})
-
-	t.Run("returns the value in megabyte", func(t *testing.T) {
-		v := MemSizeFactor64(128 * 1024 * 1024)
-		assert.Equal(t, 128, v.Megabytes())
 	})
 }
